requirements.txt
numpy
pandas
Pillow
opencv-python-headless
streamlit
tensorflow==2.15.0